backend route protection

+++ server/app.py
@@ -31,13 +31,14 @@ class Portfolios(Resource):
     def post(self):
         data = request.get_json()
 
-        try:
-            new_portfolio = Portfolio(**data)
-            db.session.add(new_portfolio)
-            db.session.commit()
-        except Exception as e:
-            db.session.rollback()
-            return make_response({"error": str(e)}, 400)
+        if session['user_id']:
+            try:
+                new_portfolio = Portfolio(**data)
+                db.session.add(new_portfolio)
+                db.session.commit()
+            except Exception as e:
+                db.session.rollback()
+                return make_response({"error": str(e)}, 400)
 
 
 class Products(Resource):
@@ -51,20 +52,20 @@ class Products(Resource):
         data = request.get_json()
         selected_tag_ids = data.get("tags", [])
         tags = Tag.query.filter(Tag.id.in_(selected_tag_ids)).all()
-
-        try:
-            new_product = Product(
-                name=data["name"],
-                description=data["description"],
-                price=data["price"],
-                tags=tags
-            )
-            db.session.add(new_product)
-            db.session.commit()
-            return make_response(new_product.to_dict(), 201)
-        except Exception as e:
-            db.session.rollback()
-            return make_response({"error": str(e)}, 400)
+        if session['user_id']:
+            try:
+                new_product = Product(
+                    name=data["name"],
+                    description=data["description"],
+                    price=data["price"],
+                    tags=tags
+                )
+                db.session.add(new_product)
+                db.session.commit()
+                return make_response(new_product.to_dict(), 201)
+            except Exception as e:
+                db.session.rollback()
+                return make_response({"error": str(e)}, 400)
 
 
 class PortfolioById(Resource):
@@ -81,7 +82,7 @@ class PortfolioById(Resource):
     def delete(self, id):
         try:
             portfolio = db.session.get(Portfolio, id)
-            if portfolio:
+            if portfolio and session['user_id']:
                 db.session.delete(portfolio)
                 db.session.commit()
                 return make_response({}, 204)
@@ -118,7 +119,7 @@ class ProductById(Resource):
     def delete(self, id):
         try:
             product = db.session.get(Product, id)
-            if product:
+            if product and session['user_id']:
                 db.session.delete(product)
                 db.session.commit()
                 return make_response({}, 204)
@@ -204,7 +205,7 @@ class Photos(Resource):
             if file.filename == '':
                 return make_response({"error": "No file selected"}, 400)
 
-            if allowed_file(file.filename):
+            if allowed_file(file.filename) and session['user_id']:
                 try:
                     filename = secure_filename(file.filename)
                     file_path = os.path.join("uploads", filename)
@@ -212,8 +213,6 @@ class Photos(Resource):
 
                     data = request.form.to_dict()
                     
-                    # if data['owner_type'] == 'product':
-                    #     data['owner_id'] = len(Product.query.all()) + 1
                     
                     photo = Photo(file_path=file_path, **data)
                     db.session.add(photo)
